+++ src/main/java/roz/tan/retail/models/Stock.java
@@ -1,8 +1,20 @@
 package roz.tan.retail.models;
 
-import lombok.*;
-import jakarta.persistence.*;
-import java.math.BigDecimal;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Table;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.UniqueConstraint;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
 import java.time.LocalDateTime;
 
 @Data
@@ -11,7 +23,12 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @Entity
 @Table(name = "stock", uniqueConstraints = {
-        @UniqueConstraint(name = "uq_stock_product_warehouse", columnNames = {"product_id", "warehouse_id"})
+        @UniqueConstraint(
+                name = "uq_stock_product_warehouse",
+                columnNames = {
+                        "product_id", "warehouse_id"
+                }
+        )
 })
 public class Stock {
 
@@ -21,22 +38,46 @@ public class Stock {
     private Integer stockId;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "product_id", nullable = false)
+    @JoinColumn(
+            name = "product_id",
+            nullable = false
+    )
     private Product product;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "warehouse_id", nullable = false)
+    @JoinColumn(
+            name = "warehouse_id",
+            nullable = false
+    )
     private Warehouse warehouse;
 
-    @Column(name = "quantity", precision = 12, scale = 3, nullable = false, columnDefinition = "DECIMAL(12,3) DEFAULT 0")
-    private BigDecimal quantity;
+    @Column(
+            name = "quantity",
+            precision = 12,
+            scale = 3,
+            nullable = false,
+            columnDefinition = "DECIMAL(12,3) DEFAULT 0"
+    )
+    private double quantity;
 
-    @Column(name = "reserved_quantity", precision = 12, scale = 3, nullable = false, columnDefinition = "DECIMAL(12,3) DEFAULT 0")
-    private BigDecimal reservedQuantity;
+    @Column(
+            name = "reserved_quantity",
+            precision = 12,
+            scale = 3,
+            nullable = false,
+            columnDefinition = "DECIMAL(12,3) DEFAULT 0"
+    )
+    private double reservedQuantity;
 
-    @Column(name = "last_updated", columnDefinition = "TIMESTAMPTZ DEFAULT CURRENT_TIMESTAMP")
+    @Column(
+            name = "last_updated",
+            columnDefinition = "TIMESTAMPTZ DEFAULT CURRENT_TIMESTAMP"
+    )
     private LocalDateTime lastUpdated;
 
-    @Column(name = "created_at", columnDefinition = "TIMESTAMPTZ DEFAULT CURRENT_TIMESTAMP")
+    @Column(
+            name = "created_at",
+            columnDefinition = "TIMESTAMPTZ DEFAULT CURRENT_TIMESTAMP"
+    )
     private LocalDateTime createdAt;
 }
